new revisions

+++ app/src/main/java/kushbhati/camcode/domain/AnalysisPoint.kt
@@ -2,5 +2,5 @@ package kushbhati.camcode.domain
 
 data class AnalysisPoint(
     val timeStamp: Long,
-    val value: Double
+    val value: Long
 )
